refactor: splitted tests
