+++ __test__/http-decorators.test.ts
@@ -1,0 +1,721 @@
+import { expect, test } from "bun:test";
+import { Route } from "../src";
+import {
+  Controller,
+  Get,
+  Before,
+  After,
+  Post,
+  Put,
+  Delete,
+  Head,
+  Patch,
+  Options,
+  All,
+} from "../src/decorators";
+import { HttpMethod } from "../src/utils";
+import "reflect-metadata";
+
+test("Get Decorator", () => {
+  @Controller("/")
+  class Test {
+    @Get()
+    public get() {}
+
+    @Get("/test")
+    public get1() {}
+
+    @Get("/test/:id")
+    public get2() {}
+
+    @Get("/test/:id/:name")
+    @Before(() => {})
+    public get3() {}
+
+    @Get("/test/:id/:name")
+    @After(() => {})
+    public get4() {}
+
+    @Get("/test/:id/:name")
+    @Before(() => {})
+    @After(() => {})
+    public get5() {}
+  }
+
+  const test = new Test();
+
+  const expectedValues = [
+    {
+      path: "/get",
+      fnName: "get",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test",
+      fnName: "get1",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id",
+      fnName: "get2",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "get3",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 1,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "get4",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "get5",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 1,
+    },
+  ];
+  const routes = Reflect.getMetadata("routes", test) as Route[];
+  expect(typeof routes).toBe("object");
+  expect(routes.length).toBe(expectedValues.length);
+
+  for (let i = 0; i < routes.length; i++) {
+    const expected = expectedValues[i];
+    const route = routes[i];
+    expect(route.after.length).toBe(expected.afterCount);
+    expect(route.before.length).toBe(expected.beforeCount);
+    expect(route.path).toBe(expected.path);
+    expect(route.fnName).toBe(expected.fnName);
+    expect(route.pathHasParams).toBe(expected.hasParams);
+    expect(route.method).toBe(HttpMethod.GET);
+  }
+});
+
+test("Post Decorator", () => {
+  @Controller("/")
+  class Test {
+    @Post()
+    public post() {}
+
+    @Post("/test")
+    public post1() {}
+
+    @Post("/test/:id")
+    public post2() {}
+
+    @Post("/test/:id/:name")
+    @Before(() => {})
+    public post3() {}
+
+    @Post("/test/:id/:name")
+    @After(() => {})
+    public post4() {}
+
+    @Post("/test/:id/:name")
+    @Before(() => {})
+    @After(() => {})
+    public post5() {}
+  }
+
+  const test = new Test();
+
+  const expectedValues = [
+    {
+      path: "/post",
+      fnName: "post",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test",
+      fnName: "post1",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id",
+      fnName: "post2",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "post3",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 1,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "post4",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "post5",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 1,
+    },
+  ];
+  const routes = Reflect.getMetadata("routes", test) as Route[];
+  expect(typeof routes).toBe("object");
+  expect(routes.length).toBe(expectedValues.length);
+
+  for (let i = 0; i < routes.length; i++) {
+    const expected = expectedValues[i];
+    const route = routes[i];
+    expect(route.after.length).toBe(expected.afterCount);
+    expect(route.before.length).toBe(expected.beforeCount);
+    expect(route.path).toBe(expected.path);
+    expect(route.fnName).toBe(expected.fnName);
+    expect(route.pathHasParams).toBe(expected.hasParams);
+    expect(route.method).toBe(HttpMethod.POST);
+  }
+});
+
+test("Put Decorator", () => {
+  @Controller("/")
+  class Test {
+    @Put()
+    public put() {}
+
+    @Put("/test")
+    public put1() {}
+
+    @Put("/test/:id")
+    public put2() {}
+
+    @Put("/test/:id/:name")
+    @Before(() => {})
+    public put3() {}
+
+    @Put("/test/:id/:name")
+    @After(() => {})
+    public put4() {}
+
+    @Put("/test/:id/:name")
+    @Before(() => {})
+    @After(() => {})
+    public put5() {}
+  }
+
+  const test = new Test();
+
+  const expectedValues = [
+    {
+      path: "/put",
+      fnName: "put",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test",
+      fnName: "put1",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id",
+      fnName: "put2",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "put3",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 1,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "put4",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "put5",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 1,
+    },
+  ];
+  const routes = Reflect.getMetadata("routes", test) as Route[];
+  expect(typeof routes).toBe("object");
+  expect(routes.length).toBe(expectedValues.length);
+
+  for (let i = 0; i < routes.length; i++) {
+    const expected = expectedValues[i];
+    const route = routes[i];
+    expect(route.after.length).toBe(expected.afterCount);
+    expect(route.before.length).toBe(expected.beforeCount);
+    expect(route.path).toBe(expected.path);
+    expect(route.fnName).toBe(expected.fnName);
+    expect(route.pathHasParams).toBe(expected.hasParams);
+    expect(route.method).toBe(HttpMethod.PUT);
+  }
+});
+
+test("Delete Decorator", () => {
+  @Controller("/")
+  class Test {
+    @Delete()
+    public delete() {}
+
+    @Delete("/test")
+    public delete1() {}
+
+    @Delete("/test/:id")
+    public delete2() {}
+
+    @Delete("/test/:id/:name")
+    @Before(() => {})
+    public delete3() {}
+
+    @Delete("/test/:id/:name")
+    @After(() => {})
+    public delete4() {}
+
+    @Delete("/test/:id/:name")
+    @Before(() => {})
+    @After(() => {})
+    public delete5() {}
+  }
+
+  const test = new Test();
+
+  const expectedValues = [
+    {
+      path: "/delete",
+      fnName: "delete",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test",
+      fnName: "delete1",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id",
+      fnName: "delete2",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "delete3",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 1,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "delete4",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "delete5",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 1,
+    },
+  ];
+  const routes = Reflect.getMetadata("routes", test) as Route[];
+  expect(typeof routes).toBe("object");
+  expect(routes.length).toBe(expectedValues.length);
+
+  for (let i = 0; i < routes.length; i++) {
+    const expected = expectedValues[i];
+    const route = routes[i];
+    expect(route.after.length).toBe(expected.afterCount);
+    expect(route.before.length).toBe(expected.beforeCount);
+    expect(route.path).toBe(expected.path);
+    expect(route.fnName).toBe(expected.fnName);
+    expect(route.pathHasParams).toBe(expected.hasParams);
+    expect(route.method).toBe(HttpMethod.DELETE);
+  }
+});
+
+test("Head Decorator", () => {
+  @Controller("/")
+  class Test {
+    @Head()
+    public head() {}
+
+    @Head("/test")
+    public head1() {}
+
+    @Head("/test/:id")
+    public head2() {}
+
+    @Head("/test/:id/:name")
+    @Before(() => {})
+    public head3() {}
+
+    @Head("/test/:id/:name")
+    @After(() => {})
+    public head4() {}
+
+    @Head("/test/:id/:name")
+    @Before(() => {})
+    @After(() => {})
+    public head5() {}
+  }
+
+  const test = new Test();
+
+  const expectedValues = [
+    {
+      path: "/head",
+      fnName: "head",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test",
+      fnName: "head1",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id",
+      fnName: "head2",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "head3",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 1,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "head4",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "head5",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 1,
+    },
+  ];
+  const routes = Reflect.getMetadata("routes", test) as Route[];
+  expect(typeof routes).toBe("object");
+  expect(routes.length).toBe(expectedValues.length);
+
+  for (let i = 0; i < routes.length; i++) {
+    const expected = expectedValues[i];
+    const route = routes[i];
+    expect(route.after.length).toBe(expected.afterCount);
+    expect(route.before.length).toBe(expected.beforeCount);
+    expect(route.path).toBe(expected.path);
+    expect(route.fnName).toBe(expected.fnName);
+    expect(route.pathHasParams).toBe(expected.hasParams);
+    expect(route.method).toBe(HttpMethod.HEAD);
+  }
+});
+
+test("Patch Decorator", () => {
+  @Controller("/")
+  class Test {
+    @Patch()
+    public patch() {}
+
+    @Patch("/test")
+    public patch1() {}
+
+    @Patch("/test/:id")
+    public patch2() {}
+
+    @Patch("/test/:id/:name")
+    @Before(() => {})
+    public patch3() {}
+
+    @Patch("/test/:id/:name")
+    @After(() => {})
+    public patch4() {}
+
+    @Patch("/test/:id/:name")
+    @Before(() => {})
+    @After(() => {})
+    public patch5() {}
+  }
+
+  const test = new Test();
+
+  const expectedValues = [
+    {
+      path: "/patch",
+      fnName: "patch",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test",
+      fnName: "patch1",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id",
+      fnName: "patch2",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "patch3",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 1,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "patch4",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "patch5",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 1,
+    },
+  ];
+  const routes = Reflect.getMetadata("routes", test) as Route[];
+  expect(typeof routes).toBe("object");
+  expect(routes.length).toBe(expectedValues.length);
+
+  for (let i = 0; i < routes.length; i++) {
+    const expected = expectedValues[i];
+    const route = routes[i];
+    expect(route.after.length).toBe(expected.afterCount);
+    expect(route.before.length).toBe(expected.beforeCount);
+    expect(route.path).toBe(expected.path);
+    expect(route.fnName).toBe(expected.fnName);
+    expect(route.pathHasParams).toBe(expected.hasParams);
+    expect(route.method).toBe(HttpMethod.PATCH);
+  }
+});
+
+test("Options Decorator", () => {
+  @Controller("/")
+  class Test {
+    @Options()
+    public options() {}
+
+    @Options("/test")
+    public options1() {}
+
+    @Options("/test/:id")
+    public options2() {}
+
+    @Options("/test/:id/:name")
+    @Before(() => {})
+    public options3() {}
+
+    @Options("/test/:id/:name")
+    @After(() => {})
+    public options4() {}
+
+    @Options("/test/:id/:name")
+    @Before(() => {})
+    @After(() => {})
+    public options5() {}
+  }
+
+  const test = new Test();
+
+  const expectedValues = [
+    {
+      path: "/options",
+      fnName: "options",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test",
+      fnName: "options1",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id",
+      fnName: "options2",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "options3",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 1,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "options4",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "options5",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 1,
+    },
+  ];
+  const routes = Reflect.getMetadata("routes", test) as Route[];
+  expect(typeof routes).toBe("object");
+  expect(routes.length).toBe(expectedValues.length);
+
+  for (let i = 0; i < routes.length; i++) {
+    const expected = expectedValues[i];
+    const route = routes[i];
+    expect(route.after.length).toBe(expected.afterCount);
+    expect(route.before.length).toBe(expected.beforeCount);
+    expect(route.path).toBe(expected.path);
+    expect(route.fnName).toBe(expected.fnName);
+    expect(route.pathHasParams).toBe(expected.hasParams);
+    expect(route.method).toBe(HttpMethod.OPTIONS);
+  }
+});
+
+test("All Decorator", () => {
+  @Controller("/")
+  class Test {
+    @All()
+    public all() {}
+
+    @All("/test")
+    public all1() {}
+
+    @All("/test/:id")
+    public all2() {}
+
+    @All("/test/:id/:name")
+    @Before(() => {})
+    public all3() {}
+
+    @All("/test/:id/:name")
+    @After(() => {})
+    public all4() {}
+
+    @All("/test/:id/:name")
+    @Before(() => {})
+    @After(() => {})
+    public all5() {}
+  }
+
+  const test = new Test();
+
+  const expectedValues = [
+    {
+      path: "/all",
+      fnName: "all",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test",
+      fnName: "all1",
+      hasParams: false,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id",
+      fnName: "all2",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "all3",
+      hasParams: true,
+      afterCount: 0,
+      beforeCount: 1,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "all4",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 0,
+    },
+    {
+      path: "/test/:id/:name",
+      fnName: "all5",
+      hasParams: true,
+      afterCount: 1,
+      beforeCount: 1,
+    },
+  ];
+  const routes = Reflect.getMetadata("routes", test) as Route[];
+  expect(typeof routes).toBe("object");
+  expect(routes.length).toBe(expectedValues.length);
+
+  for (let i = 0; i < routes.length; i++) {
+    const expected = expectedValues[i];
+    const route = routes[i];
+    expect(route.after.length).toBe(expected.afterCount);
+    expect(route.before.length).toBe(expected.beforeCount);
+    expect(route.path).toBe(expected.path);
+    expect(route.fnName).toBe(expected.fnName);
+    expect(route.pathHasParams).toBe(expected.hasParams);
+    expect(route.method).toBe(HttpMethod.ALL);
+  }
+});
