+++ __test__/ikari.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { expect, test, describe, jest } from "bun:test";
+import { expect, test, describe } from "bun:test";
 import {
   HttpMethod,
   NotFound,
@@ -28,7 +28,7 @@ import "reflect-metadata";
 import { Config, Context, Route, Serve } from "../src";
 import { unlinkSync } from "node:fs";
 import { Errorlike } from "bun";
-import { helmet } from "../src/middlewares/helmet";
+import { createContextMock } from "./helper";
 
 test("createPath", () => {
   expect(createPath("test")).toBe("/test");
@@ -126,622 +126,6 @@ describe("ServeValidator", () => {
       }).validate();
     }).toThrow("Strict must be a boolean");
   });
-});
-
-test("Get Decorator", () => {
-  @Controller("/")
-  class Test {
-    @Get()
-    public get() {}
-
-    @Get("/test")
-    public get1() {}
-
-    @Get("/test/:id")
-    public get2() {}
-
-    @Get("/test/:id/:name")
-    @Before(() => {})
-    public get3() {}
-
-    @Get("/test/:id/:name")
-    @After(() => {})
-    public get4() {}
-
-    @Get("/test/:id/:name")
-    @Before(() => {})
-    @After(() => {})
-    public get5() {}
-  }
-
-  const test = new Test();
-
-  const expectedValues = [
-    {
-      path: "/get",
-      fnName: "get",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test",
-      fnName: "get1",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id",
-      fnName: "get2",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "get3",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 1,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "get4",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "get5",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 1,
-    },
-  ];
-  const routes = Reflect.getMetadata("routes", test) as Route[];
-  expect(typeof routes).toBe("object");
-  expect(routes.length).toBe(expectedValues.length);
-
-  for (let i = 0; i < routes.length; i++) {
-    const expected = expectedValues[i];
-    const route = routes[i];
-    expect(route.after.length).toBe(expected.afterCount);
-    expect(route.before.length).toBe(expected.beforeCount);
-    expect(route.path).toBe(expected.path);
-    expect(route.fnName).toBe(expected.fnName);
-    expect(route.pathHasParams).toBe(expected.hasParams);
-    expect(route.method).toBe(HttpMethod.GET);
-  }
-});
-
-test("Post Decorator", () => {
-  @Controller("/")
-  class Test {
-    @Post()
-    public post() {}
-
-    @Post("/test")
-    public post1() {}
-
-    @Post("/test/:id")
-    public post2() {}
-
-    @Post("/test/:id/:name")
-    @Before(() => {})
-    public post3() {}
-
-    @Post("/test/:id/:name")
-    @After(() => {})
-    public post4() {}
-
-    @Post("/test/:id/:name")
-    @Before(() => {})
-    @After(() => {})
-    public post5() {}
-  }
-
-  const test = new Test();
-
-  const expectedValues = [
-    {
-      path: "/post",
-      fnName: "post",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test",
-      fnName: "post1",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id",
-      fnName: "post2",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "post3",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 1,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "post4",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "post5",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 1,
-    },
-  ];
-  const routes = Reflect.getMetadata("routes", test) as Route[];
-  expect(typeof routes).toBe("object");
-  expect(routes.length).toBe(expectedValues.length);
-
-  for (let i = 0; i < routes.length; i++) {
-    const expected = expectedValues[i];
-    const route = routes[i];
-    expect(route.after.length).toBe(expected.afterCount);
-    expect(route.before.length).toBe(expected.beforeCount);
-    expect(route.path).toBe(expected.path);
-    expect(route.fnName).toBe(expected.fnName);
-    expect(route.pathHasParams).toBe(expected.hasParams);
-    expect(route.method).toBe(HttpMethod.POST);
-  }
-});
-
-test("Put Decorator", () => {
-  @Controller("/")
-  class Test {
-    @Put()
-    public put() {}
-
-    @Put("/test")
-    public put1() {}
-
-    @Put("/test/:id")
-    public put2() {}
-
-    @Put("/test/:id/:name")
-    @Before(() => {})
-    public put3() {}
-
-    @Put("/test/:id/:name")
-    @After(() => {})
-    public put4() {}
-
-    @Put("/test/:id/:name")
-    @Before(() => {})
-    @After(() => {})
-    public put5() {}
-  }
-
-  const test = new Test();
-
-  const expectedValues = [
-    {
-      path: "/put",
-      fnName: "put",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test",
-      fnName: "put1",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id",
-      fnName: "put2",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "put3",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 1,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "put4",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "put5",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 1,
-    },
-  ];
-  const routes = Reflect.getMetadata("routes", test) as Route[];
-  expect(typeof routes).toBe("object");
-  expect(routes.length).toBe(expectedValues.length);
-
-  for (let i = 0; i < routes.length; i++) {
-    const expected = expectedValues[i];
-    const route = routes[i];
-    expect(route.after.length).toBe(expected.afterCount);
-    expect(route.before.length).toBe(expected.beforeCount);
-    expect(route.path).toBe(expected.path);
-    expect(route.fnName).toBe(expected.fnName);
-    expect(route.pathHasParams).toBe(expected.hasParams);
-    expect(route.method).toBe(HttpMethod.PUT);
-  }
-});
-
-test("Delete Decorator", () => {
-  @Controller("/")
-  class Test {
-    @Delete()
-    public delete() {}
-
-    @Delete("/test")
-    public delete1() {}
-
-    @Delete("/test/:id")
-    public delete2() {}
-
-    @Delete("/test/:id/:name")
-    @Before(() => {})
-    public delete3() {}
-
-    @Delete("/test/:id/:name")
-    @After(() => {})
-    public delete4() {}
-
-    @Delete("/test/:id/:name")
-    @Before(() => {})
-    @After(() => {})
-    public delete5() {}
-  }
-
-  const test = new Test();
-
-  const expectedValues = [
-    {
-      path: "/delete",
-      fnName: "delete",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test",
-      fnName: "delete1",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id",
-      fnName: "delete2",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "delete3",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 1,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "delete4",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "delete5",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 1,
-    },
-  ];
-  const routes = Reflect.getMetadata("routes", test) as Route[];
-  expect(typeof routes).toBe("object");
-  expect(routes.length).toBe(expectedValues.length);
-
-  for (let i = 0; i < routes.length; i++) {
-    const expected = expectedValues[i];
-    const route = routes[i];
-    expect(route.after.length).toBe(expected.afterCount);
-    expect(route.before.length).toBe(expected.beforeCount);
-    expect(route.path).toBe(expected.path);
-    expect(route.fnName).toBe(expected.fnName);
-    expect(route.pathHasParams).toBe(expected.hasParams);
-    expect(route.method).toBe(HttpMethod.DELETE);
-  }
-});
-
-test("Head Decorator", () => {
-  @Controller("/")
-  class Test {
-    @Head()
-    public head() {}
-
-    @Head("/test")
-    public head1() {}
-
-    @Head("/test/:id")
-    public head2() {}
-
-    @Head("/test/:id/:name")
-    @Before(() => {})
-    public head3() {}
-
-    @Head("/test/:id/:name")
-    @After(() => {})
-    public head4() {}
-
-    @Head("/test/:id/:name")
-    @Before(() => {})
-    @After(() => {})
-    public head5() {}
-  }
-
-  const test = new Test();
-
-  const expectedValues = [
-    {
-      path: "/head",
-      fnName: "head",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test",
-      fnName: "head1",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id",
-      fnName: "head2",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "head3",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 1,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "head4",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "head5",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 1,
-    },
-  ];
-  const routes = Reflect.getMetadata("routes", test) as Route[];
-  expect(typeof routes).toBe("object");
-  expect(routes.length).toBe(expectedValues.length);
-
-  for (let i = 0; i < routes.length; i++) {
-    const expected = expectedValues[i];
-    const route = routes[i];
-    expect(route.after.length).toBe(expected.afterCount);
-    expect(route.before.length).toBe(expected.beforeCount);
-    expect(route.path).toBe(expected.path);
-    expect(route.fnName).toBe(expected.fnName);
-    expect(route.pathHasParams).toBe(expected.hasParams);
-    expect(route.method).toBe(HttpMethod.HEAD);
-  }
-});
-
-test("Patch Decorator", () => {
-  @Controller("/")
-  class Test {
-    @Patch()
-    public patch() {}
-
-    @Patch("/test")
-    public patch1() {}
-
-    @Patch("/test/:id")
-    public patch2() {}
-
-    @Patch("/test/:id/:name")
-    @Before(() => {})
-    public patch3() {}
-
-    @Patch("/test/:id/:name")
-    @After(() => {})
-    public patch4() {}
-
-    @Patch("/test/:id/:name")
-    @Before(() => {})
-    @After(() => {})
-    public patch5() {}
-  }
-
-  const test = new Test();
-
-  const expectedValues = [
-    {
-      path: "/patch",
-      fnName: "patch",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test",
-      fnName: "patch1",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id",
-      fnName: "patch2",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "patch3",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 1,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "patch4",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "patch5",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 1,
-    },
-  ];
-  const routes = Reflect.getMetadata("routes", test) as Route[];
-  expect(typeof routes).toBe("object");
-  expect(routes.length).toBe(expectedValues.length);
-
-  for (let i = 0; i < routes.length; i++) {
-    const expected = expectedValues[i];
-    const route = routes[i];
-    expect(route.after.length).toBe(expected.afterCount);
-    expect(route.before.length).toBe(expected.beforeCount);
-    expect(route.path).toBe(expected.path);
-    expect(route.fnName).toBe(expected.fnName);
-    expect(route.pathHasParams).toBe(expected.hasParams);
-    expect(route.method).toBe(HttpMethod.PATCH);
-  }
-});
-
-test("Options Decorator", () => {
-  @Controller("/")
-  class Test {
-    @Options()
-    public options() {}
-
-    @Options("/test")
-    public options1() {}
-
-    @Options("/test/:id")
-    public options2() {}
-
-    @Options("/test/:id/:name")
-    @Before(() => {})
-    public options3() {}
-
-    @Options("/test/:id/:name")
-    @After(() => {})
-    public options4() {}
-
-    @Options("/test/:id/:name")
-    @Before(() => {})
-    @After(() => {})
-    public options5() {}
-  }
-
-  const test = new Test();
-
-  const expectedValues = [
-    {
-      path: "/options",
-      fnName: "options",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test",
-      fnName: "options1",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id",
-      fnName: "options2",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "options3",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 1,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "options4",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "options5",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 1,
-    },
-  ];
-  const routes = Reflect.getMetadata("routes", test) as Route[];
-  expect(typeof routes).toBe("object");
-  expect(routes.length).toBe(expectedValues.length);
-
-  for (let i = 0; i < routes.length; i++) {
-    const expected = expectedValues[i];
-    const route = routes[i];
-    expect(route.after.length).toBe(expected.afterCount);
-    expect(route.before.length).toBe(expected.beforeCount);
-    expect(route.path).toBe(expected.path);
-    expect(route.fnName).toBe(expected.fnName);
-    expect(route.pathHasParams).toBe(expected.hasParams);
-    expect(route.method).toBe(HttpMethod.OPTIONS);
-  }
 });
 
 test("Controller Decorator", () => {
@@ -1041,94 +425,6 @@ test("Controller Decorator", () => {
   }
 });
 
-test("All Decorator", () => {
-  @Controller("/")
-  class Test {
-    @All()
-    public all() {}
-
-    @All("/test")
-    public all1() {}
-
-    @All("/test/:id")
-    public all2() {}
-
-    @All("/test/:id/:name")
-    @Before(() => {})
-    public all3() {}
-
-    @All("/test/:id/:name")
-    @After(() => {})
-    public all4() {}
-
-    @All("/test/:id/:name")
-    @Before(() => {})
-    @After(() => {})
-    public all5() {}
-  }
-
-  const test = new Test();
-
-  const expectedValues = [
-    {
-      path: "/all",
-      fnName: "all",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test",
-      fnName: "all1",
-      hasParams: false,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id",
-      fnName: "all2",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "all3",
-      hasParams: true,
-      afterCount: 0,
-      beforeCount: 1,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "all4",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 0,
-    },
-    {
-      path: "/test/:id/:name",
-      fnName: "all5",
-      hasParams: true,
-      afterCount: 1,
-      beforeCount: 1,
-    },
-  ];
-  const routes = Reflect.getMetadata("routes", test) as Route[];
-  expect(typeof routes).toBe("object");
-  expect(routes.length).toBe(expectedValues.length);
-
-  for (let i = 0; i < routes.length; i++) {
-    const expected = expectedValues[i];
-    const route = routes[i];
-    expect(route.after.length).toBe(expected.afterCount);
-    expect(route.before.length).toBe(expected.beforeCount);
-    expect(route.path).toBe(expected.path);
-    expect(route.fnName).toBe(expected.fnName);
-    expect(route.pathHasParams).toBe(expected.hasParams);
-    expect(route.method).toBe(HttpMethod.ALL);
-  }
-});
-
 test("Controller Decorator with prefix", () => {
   @Controller("/test")
   class Test {
@@ -1220,7 +516,7 @@ test("After Decorator", () => {
   expect(route.after[1]).toBe(after1);
 });
 
-test("Context ", async () => {
+test("Context", async () => {
   @Controller("/")
   class Test {
     @Get("/get")
@@ -3461,36 +2757,6 @@ describe("Controller Type", async () => {
   });
 });
 
-const createContextMock = (method: HttpMethod) => {
-  const statusMock = jest.fn();
-  const jsonMock = jest.fn();
-  const getResWithoutBodyMock = jest.fn();
-  const resMock = {} as Response;
-  const context = {
-    method,
-    status: (param: number) => {
-      statusMock(param);
-      return context;
-    },
-    getResWithoutBody: () => {
-      getResWithoutBodyMock();
-      return context;
-    },
-    json: (param: any, status: number) => {
-      jsonMock(param, status);
-      return context;
-    },
-    res: resMock,
-  } as unknown as Context;
-
-  return {
-    context,
-    statusMock,
-    jsonMock,
-    getResWithoutBodyMock,
-    resMock,
-  };
-};
 
 describe("NotFound function", async () => {
   test("when NotFound called with HEAD it returns 404 without body.", async () => {
@@ -4839,170 +4105,3 @@ describe("Request Cycle", () => {
     serve.stop();
   });
 });
-
-describe("Helmet", () => {
-  test("Helmet should be work", async () => {
-    @Controller("/test")
-    class Test {
-      @Get("/test")
-      public get(ctx: Context) {
-        return ctx.json({ fn: "get", method: ctx.method });
-      }
-    }
-
-    const config: Config = {
-      controllers: [Test],
-      disableStartupMessage: true,
-      middlewares: [
-        helmet({
-          ContentSecurityPolicy: "default-src 'self'",
-          ContentTypeNosniff: "none",
-          CrossOriginEmbedderPolicy: "require-corp",
-          CrossOriginOpenerPolicy: "same-origin",
-          CrossOriginResourcePolicy: "same-origin",
-          CSPReportOnly: false,
-          XFrameOptions: "SAMEORIGIN",
-          ReferrerPolicy: "same-origin",
-          XSSProtection: "1",
-        }),
-      ],
-      serveOptions: {
-        port: 0,
-      },
-    };
-
-    const serve = Serve(config);
-    const expectedValues = [
-      {
-        path: "/test/test",
-        method: HttpMethod.GET,
-        statusCode: StatusCode.OK,
-        headers: {
-          "content-security-policy": "default-src 'self'",
-          "x-content-type-options": "none",
-          "cross-origin-embedder-policy": "require-corp",
-          "cross-origin-opener-policy": "same-origin",
-          "cross-origin-resource-policy": "same-origin",
-          "x-frame-options": "SAMEORIGIN",
-          "referrer-policy": "same-origin",
-          "x-xss-protection": "1",
-        },
-      },
-    ];
-
-    for (const expected of expectedValues) {
-      const res = await fetch(
-        serve.hostname + ":" + serve.port + expected.path,
-        {
-          method: expected.method,
-        }
-      );
-      expect(res.status).toBe(expected.statusCode);
-      Object.entries(expected.headers).forEach(([key, value]) => {
-        expect(res.headers.get(key)?.replace(/\s/g, "")).toBe(
-          value.replace(/\s/g, "")
-        );
-      });
-    }
-
-    serve.stop();
-  });
-
-  test("Helmet should be work with default options", async () => {
-    @Controller("/test")
-    class Test {
-      @Get("/test")
-      public get(ctx: Context) {
-        return ctx.json({ fn: "get", method: ctx.method });
-      }
-    }
-
-    const config: Config = {
-      controllers: [Test],
-      disableStartupMessage: true,
-      middlewares: [helmet()],
-      serveOptions: {
-        port: 0,
-      },
-    };
-
-    const serve = Serve(config);
-    const expectedValues = [
-      {
-        path: "/test/test",
-        method: HttpMethod.GET,
-        statusCode: StatusCode.OK,
-        headers: {
-          "x-content-type-options": "nosniff",
-          "x-frame-options": "SAMEORIGIN",
-          "x-xss-protection": "0",
-        },
-      },
-    ];
-
-    for (const expected of expectedValues) {
-      const res = await fetch(
-        serve.hostname + ":" + serve.port + expected.path,
-        {
-          method: expected.method,
-        }
-      );
-      expect(res.status).toBe(expected.statusCode);
-      Object.entries(expected.headers).forEach(([key, value]) => {
-        expect(res.headers.get(key)?.replace(/\s/g, "")).toBe(
-          value.replace(/\s/g, "")
-        );
-      });
-    }
-
-    serve.stop();
-  });
-
-  test("Helmet should be work with default options and CSPReportOnly true", async () => {
-    @Controller("/test")
-    class Test {
-      @Get("/test")
-      public get(ctx: Context) {
-        return ctx.json({ fn: "get", method: ctx.method });
-      }
-    }
-
-    const config: Config = {
-      controllers: [Test],
-      disableStartupMessage: true,
-      middlewares: [helmet({ CSPReportOnly: true })],
-      serveOptions: {
-        port: 0,
-      },
-    };
-
-    const serve = Serve(config);
-    const expectedValues = [
-      {
-        path: "/test/test",
-        method: HttpMethod.GET,
-        statusCode: StatusCode.OK,
-        headers: {},
-      },
-    ];
-
-    for (const expected of expectedValues) {
-      const res = await fetch(
-        serve.hostname + ":" + serve.port + expected.path,
-        {
-          method: expected.method,
-        }
-      );
-      expect(res.status).toBe(expected.statusCode);
-      expect(res.headers.get("content-security-policy-report-only")).toBeNull();
-    }
-
-    serve.stop();
-  });
-});
-
-// TODO middleware type check
-// TODO locals delete and clear test
-// TODO CORS test
-// TODO add requestId test
-// TODO add logger test
